Seperate food item definition from hunger plugin
