+++ _Hunger.cs
@@ -1,5 +1,3 @@
-//pluginref __itemsystem.dll
-//pluginref __constants.dll
 using System;
 using System.Collections.Generic;
 using MCGalaxy;
@@ -9,29 +7,7 @@ using MCGalaxy.Maths;
 using MCGalaxy.SQL;
 using MCGalaxy.Tasks;
 using BlockID = System.UInt16;
-using ProjectCommunity.Items;
-using ProjectCommunity.Items.Food;
 using ProjectCommunity;
-namespace ProjectCommunity.Items.Food
-{
-    public class FoodBase : ItemBaseConsumeable
-    {
-        int HungerReplenishAmount = 1;
-        public override void OnUse(Player p, ushort x, ushort y, ushort z, byte entity)
-        {
-            Hunger.AddHunger(p, HungerReplenishAmount);
-        }
-        public override bool CanUse(Player p)
-        {
-            return base.CanUse(p) && Hunger.GetHunger(p) < Hunger.MaxHunger;
-        }
-        public FoodBase(int replenishment=1) : base()
-        {
-            this.HungerReplenishAmount=replenishment;
-        }
-    }
-
-}
 
 namespace ProjectCommunity
 {
@@ -42,7 +18,6 @@ namespace ProjectCommunity
         public override string creator { get { return "morgana"; } }
 
         public const int MaxHunger = 10;
-        public const float StarvingWalkSpeed = 0.8f;
 
         private SchedulerTask hungerTask;
 
@@ -60,23 +35,6 @@ namespace ProjectCommunity
             Server.MainScheduler.QueueRepeat(HungerTick, null, TimeSpan.FromMilliseconds(500));
 
             OnJoinedLevelEvent.Register(PlayerJoinedLevel, Priority.Normal);
-
-            ItemSystem.RegisterItem(39, new FoodBase(2)); // Temp mushroom test
-
-            ItemSystem.RegisterItem(BlockConstants.Carrot, new FoodBase(2));
-            ItemSystem.RegisterItem(BlockConstants.Beet, new FoodBase(2));
-            ItemSystem.RegisterItem(BlockConstants.Potato, new FoodBase(2));
-
-            ItemSystem.RegisterItem(BlockConstants.BakedPotato, new FoodBase(4));
-            ItemSystem.RegisterItem(BlockConstants.Cake, new FoodBase(4));
-            ItemSystem.RegisterItem(BlockConstants.Bread, new FoodBase(3));
-            ItemSystem.RegisterItem(BlockConstants.Chicken, new FoodBase(2));
-            ItemSystem.RegisterItem(BlockConstants.Steak, new FoodBase(2));
-            ItemSystem.RegisterItem(BlockConstants.Fish, new FoodBase(4));
-            ItemSystem.RegisterItem(BlockConstants.Cookie, new FoodBase(3));
-            ItemSystem.RegisterItem(BlockConstants.Grape, new FoodBase(1));
-            ItemSystem.RegisterItem(BlockConstants.Soup, new FoodBase(4));
-
         }
 
         public override void Unload(bool shutdown)
@@ -84,22 +42,6 @@ namespace ProjectCommunity
             Server.MainScheduler.Cancel(hungerTask);
 
             OnJoinedLevelEvent.Unregister(PlayerJoinedLevel);
-
-            ItemSystem.UnregisterItem(39);
-
-            ItemSystem.UnregisterItem(BlockConstants.Carrot);
-            ItemSystem.UnregisterItem(BlockConstants.Beet);
-            ItemSystem.UnregisterItem(BlockConstants.Potato);
-
-            ItemSystem.UnregisterItem(BlockConstants.BakedPotato);
-            ItemSystem.UnregisterItem(BlockConstants.Cake);
-            ItemSystem.UnregisterItem(BlockConstants.Bread);
-            ItemSystem.UnregisterItem(BlockConstants.Chicken);
-            ItemSystem.UnregisterItem(BlockConstants.Steak);
-            ItemSystem.UnregisterItem(BlockConstants.Fish);
-            ItemSystem.UnregisterItem(BlockConstants.Cookie);
-            ItemSystem.UnregisterItem(BlockConstants.Grape);
-            ItemSystem.UnregisterItem(BlockConstants.Soup);
         }
 
         private void HungerTick(SchedulerTask task)
